feature(bank-sdk): Skonto. Amounts validation message
PP-763
