+++ bank-sdk/sdk/src/main/java/net/gini/android/bank/sdk/capture/skonto/SkontoScreenModule.kt
@@ -24,6 +24,8 @@ val skontoScreenModule = module {
             transactionDocDialogCancelAttachUseCase = get(),
             getTransactionDocShouldBeAutoAttachedUseCase = get(),
             getTransactionDocsFeatureEnabledUseCase = get(),
+            getFullAmountValidationErrorUseCase = get(),
+            getSkontoAmountValidationErrorUseCase = get(),
         )
     }
     factory {
