+++ bank-sdk/sdk/src/main/java/net/gini/android/bank/sdk/capture/skonto/SkontoScreenSideEffect.kt
@@ -1,0 +1,8 @@
+package net.gini.android.bank.sdk.capture.skonto
+
+internal sealed interface SkontoScreenSideEffect {
+    data class OpenInvoiceScreen(
+        val documentId: String,
+        val infoTextLines: List<String>,
+    ) : SkontoScreenSideEffect
+}
